--- app/src/main/java/com/practicum/playlistmaker/SearchResponse.kt
@@ -1,6 +1,0 @@
-package com.practicum.playlistmaker
-
-class SearchResponse(
-    val resultCount: Int,
-    val results: ArrayList<Track>
-)
