--- app/src/main/java/com/practicum/playlistmaker/Track.kt
@@ -1,28 +1,0 @@
-package com.practicum.playlistmaker
-
-import android.content.Context
-import android.util.TypedValue
-import android.view.LayoutInflater
-import android.view.RoundedCorner
-import android.view.View
-import android.view.ViewGroup
-import android.widget.ImageView
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import retrofit2.http.Url
-import java.net.URL
-
-data class Track(
-    val trackId:String,
-    val trackName: String,
-    val artistName: String,
-    val trackTimeMillis: Int,
-    val artworkUrl100: String,
-    val collectionName: String,
-    val releaseDate: String,
-    val primaryGenreName: String,
-    val country: String,
-    val previewUrl: String
-)
