+++ app/src/main/java/com/practicum/playlistmaker/TrackAdapter.kt
@@ -1,13 +1,12 @@
 package com.practicum.playlistmaker
 
-import android.content.SharedPreferences
 import android.os.Handler
 import android.os.Looper
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.appcompat.app.AppCompatActivity.MODE_PRIVATE
 import androidx.recyclerview.widget.RecyclerView
+import com.practicum.playlistmaker.domain.models.Track
 
 class TrackAdapter(
     private val tracks: List<Track>
