+++ app/src/main/java/com/practicum/playlistmaker/data/dto/TrackSearchResponse.kt
@@ -1,0 +1,8 @@
+package com.practicum.playlistmaker.data.dto
+
+import com.practicum.playlistmaker.domain.models.Track
+
+data class TrackSearchResponse(
+    val resultCount: Int,
+    val results: ArrayList<Track>
+)
