+++ app/src/main/java/com/practicum/playlistmaker/domain/models/Track.kt
@@ -1,0 +1,14 @@
+package com.practicum.playlistmaker.domain.models
+
+data class Track(
+    val trackId:String,
+    val trackName: String,
+    val artistName: String,
+    val trackTimeMillis: Int,
+    val artworkUrl100: String,
+    val collectionName: String,
+    val releaseDate: String,
+    val primaryGenreName: String,
+    val country: String,
+    val previewUrl: String
+)
