+++ app/src/main/java/com/practicum/playlistmaker/iTunesApi.kt
@@ -1,5 +1,6 @@
 package com.practicum.playlistmaker
 
+import com.practicum.playlistmaker.data.dto.TrackSearchResponse
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -9,6 +10,6 @@ interface iTunesApi {
     @GET("/search?entity=song")
     fun search(
         @Query("term") text: String
-    ): Call<SearchResponse>
+    ): Call<TrackSearchResponse>
 
 }
