+++ app/src/main/java/com/practicum/playlistmaker/ui/search/SearchActivity.kt
@@ -29,9 +29,9 @@ import com.practicum.playlistmaker.PREFERENCES
 import com.practicum.playlistmaker.R
 import com.practicum.playlistmaker.SEARCH_HISTORY_KEY
 import com.practicum.playlistmaker.SearchHistory
-import com.practicum.playlistmaker.SearchResponse
+import com.practicum.playlistmaker.data.dto.TrackSearchResponse
 import com.practicum.playlistmaker.SearchStatus
-import com.practicum.playlistmaker.Track
+import com.practicum.playlistmaker.domain.models.Track
 import com.practicum.playlistmaker.TrackAdapter
 import com.practicum.playlistmaker.iTunesApi
 import com.practicum.playlistmaker.ui.player.PlayerActivity
@@ -263,11 +263,11 @@ class SearchActivity : AppCompatActivity() {
 
 
     private fun search() {
-        iTunesService.search(searchEdittext.text.toString()).enqueue(object : Callback<SearchResponse> {
+        iTunesService.search(searchEdittext.text.toString()).enqueue(object : Callback<TrackSearchResponse> {
             @SuppressLint("NotifyDataSetChanged")
             override fun onResponse(
-                call: Call<SearchResponse>,
-                response: Response<SearchResponse>
+                call: Call<TrackSearchResponse>,
+                response: Response<TrackSearchResponse>
             ) {
                 tracks.clear()
 
@@ -291,7 +291,7 @@ class SearchActivity : AppCompatActivity() {
             }
 
             @SuppressLint("NotifyDataSetChanged")
-            override fun onFailure(call: Call<SearchResponse>, t: Throwable) {
+            override fun onFailure(call: Call<TrackSearchResponse>, t: Throwable) {
                 tracks.clear()
                 showStatus(SearchStatus.SOMETHING_WRONG, SOMETHING_WRONG)
             }
